Alert

+++ components/FavoriteComponent.js
@@ -31,7 +31,26 @@ class Favorites extends Component {
 				{
 					text: 'Delete',
 					type: 'delete',
-					onPress: () => this.props.deleteFavorite(item.id),
+					onPress: () => {
+						Alert.alert(
+							'Delete Favorite?',
+							'Are you sure you wish to delete the favorite dish ' +
+								item.name +
+								'?',
+							[
+								{
+									text: 'Cancel',
+									onPress: () => console.log(item.name + 'Not Deleted'),
+									style: ' cancel',
+								},
+								{
+									text: 'OK',
+									onPress: () => this.props.deleteFavorite(item.id),
+								},
+							],
+							{ cancelable: false }
+						);
+					},
 				},
 			];
 			return (
